Stop using tables for mini-map grids.
