+++ components/reports/permafrost/ReportMagtMap.vue
@@ -1,12 +1,41 @@
 <template>
-  <div class="has-text-centered has-text-weight-bold">
-    <div :id="mapID" class="permafrost-minimap"></div>
+  <div>
+    <div class="map-title map-container has-text-centered has-text-weight-bold">
+      <div>
+        <span v-if="mapModel">{{ mapModel }}<br /></span>
+        <span>{{ mapScenario }}</span>
+        <br class="narrow-br" />
+        <span>{{ mapEra }}</span>
+      </div>
+    </div>
+    <div :id="mapID" class="map"></div>
   </div>
 </template>
 
 <style lang="scss" scoped>
+.map-title {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  flex-direction: column;
+}
+@media (max-width: 899px) {
+  .map-title {
+    min-height: 90px;
+  }
+}
+@media (min-width: 900px) {
+  .map-title {
+    min-height: 60px;
+  }
+}
+@media (min-width: 900px) {
+  .narrow-br {
+    display: none;
+  }
+}
 /* CSS trick to make height same as dynamic width */
-.permafrost-minimap {
+.map {
   height: 0;
   padding-bottom: 100%;
 }
@@ -16,6 +45,14 @@
 import _ from 'lodash'
 import { mapGetters } from 'vuex'
 import { getBaseMapAndLayers, addGeoJSONtoMap } from '~/utils/maps'
+
+let models = {
+  3: 'NCAR CCSM4',
+  4: 'MRI CGCM3',
+}
+
+let scenarios = ['CRU TS 3.1', 'RCP 4.5', 'RCP 8.5']
+let eras = ['1995', '2011-2040', '2036-2065', '2061–2090', '2086–2100']
 
 export default {
   name: 'ReportMagtMap',
@@ -27,6 +64,17 @@ export default {
     }),
     mapID() {
       return this.scenario + '_' + this.model + '_' + this.era
+    },
+    mapModel() {
+      if (this.era > 0) {
+        return models[this.model]
+      }
+    },
+    mapScenario() {
+      return scenarios[this.scenario] + ', '
+    },
+    mapEra() {
+      return eras[this.era]
     },
   },
   data() {
